Adding $ operator.

+++ fgfl/inc/fgfl.h
@@ -93,7 +93,7 @@ typedef struct {
     bitset_t* set_state;
     size_t hash_state;
     trans_list_t* trans;
-    size_t final_entry;
+    int final_entry;
     int final_anchor_entry;
     bool middle;
     int group;
@@ -113,8 +113,6 @@ typedef struct {
             bool use_upper;
 
             bool beg_line;
-//          TODO
-//          bool end_line;
 
             bool use_look;
             bool skip;
